% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cfd.mm.weights.casrx}
\alias{cfd.mm.weights.casrx}
\title{Mismatch position weights for the CFD off-target algorithm (CasRx)}
\format{
Numeric vector
}
\usage{
cfd.mm.weights.casrx
}
\description{
CFD off-target scoring weights for each nucleotide position and each
    pair of DNA:RNA mismatches, for 27-mer spacer sequences.
    Position 1 refers to the spacer position adjacent to
    the PAM sequence. A higher weight indicates greater mismatch tolerance 
    by the CasRx nuclease.
}
\keyword{datasets}
