% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cfd.pam.weights.cas9}
\alias{cfd.pam.weights.cas9}
\title{PAM sequence weights for the CFD off-target algorithm (Cas9)}
\format{
Numeric vector
}
\source{
\url{https://dx.doi.org/10.1038\%2Fnbt.3437}
}
\usage{
cfd.pam.weights.cas9
}
\description{
CFD off-target PAM weights for each possible 3-mer Cas9 PAM sequences.
    A higher weight indicates greater mismatch tolerance 
    by the Cas9 nuclease.
}
\keyword{datasets}
