% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cfd.pam.weights.casrx}
\alias{cfd.pam.weights.casrx}
\title{PAM sequence weights for the CFD off-target algorithm (CasRx)}
\format{
Numeric vector
}
\usage{
cfd.pam.weights.casrx
}
\description{
CFD off-target PAM weights for each possible 1-mer CasRx PAM sequences.
    A higher weight indicates greater mismatch tolerance 
    by the CasRx nuclease.
}
\keyword{datasets}
