% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWeissmanScores.R
\name{getWeissmanScores}
\alias{getWeissmanScores}
\title{Calculate on-target sgRNA activity scores using crisprai}
\usage{
getWeissmanScores(
  tss_df,
  sgrna_df,
  verbose = FALSE,
  modality = c("CRISPRa", "CRISPRi"),
  fork = FALSE
)
}
\arguments{
\item{tss_df}{A \code{data.frame} containing transcription start site (TSS)
data for promoters. Must have these columns: \code{gene_symbol},
\code{promoter}, \code{transcripts}, \code{position}, \code{strand}, and
\code{chr}. See below for more details about \code{tss_df}.}

\item{sgrna_df}{\code{A data.frame} containing PAM start site and sequence
for each sgRNA. Must have these columns: \code{grna_id}, \code{tss_id},
\code{pam_site}, \code{strand}, and \code{spacer_19mer}. See below for
more details about \code{sgrna_df}.}

\item{verbose}{Should messages be printed to the console? Default value is
\code{TRUE}.}

\item{modality}{Which mode of perturbation is being used? Must be a
\code{string} specifying either \code{CRISPRa} or \code{CRISPRi}.}

\item{fork}{Set to \code{TRUE} to preserve changes to the R
configuration within the session.}
}
\value{
\strong{getWeissmanScore} returns a \code{data.frame} with
\code{sequence} and \code{score} columns. The Weissman score takes on a
value between 0 and 1. A higher score indicates higher sgRNA efficiency.
}
\description{
Use the Weissman scoring method to calculate on-target sgRNA
activity scores for Cas9-based CRISPR activation (CRISPRa) and CRISPR
inactivation (CRISPRi) gene perturbation studies. The Weissman algorithm
incorporates chromatin features, transcription start site, and sequence
to predict single-guide RNA (sgRNA) activity scores as these can all
influence activity in CRISPRa and CRISPRi perturbation studies. This
method currently only works for sgRNAs designed for use with Cas9 and
hg38 genome assembly.
}
\details{
\code{tss_df} details:
This must be a \code{data.frame} that contains the following columns:
* gene_symbol: HGNC/HUGO gene identifier.
* promoter: promoter ID.
* transcripts: Ensembl transcript identifier.
* position: start position of transcription start site (TSS).
* strand: strand location. Either \emph{+} or \emph{-}.
* chr: chromosome location for gene. \emph{e.g. chr19}.

\code{sgrna_df} details:
This must be a \code{data.frame} that contains the following columns:
* grna_id: unique sgRNA identifier
* tss_id: name of the transcription start site.
* pam_site: position of the \strong{N} in the \emph{NGG} PAM sequence.
* strand: strand location. Either \emph{+} or \emph{-}.
* spacer_19mer: sgRNA spacer sequence.
}
\examples{
\donttest{
tss.data <- read.table("tss_file.txt", header=TRUE, sep="\t")
sgrna.data <- read.table("sgrna_file.txt", header=TRUE, sep="\t")
modality <- "CRISPRa"
results <- getWeissmanScores(tss_df = tss.data,
                             sgrna_df = sgrna.data,
                             modality = modality)
}

}
\references{
Horlbeck et al. Compact and highly active next-generation libraries for
CRISPR-mediated gene repression and activation
eLife 2016;5:e19760.
\url{https://doi.org/10.7554/eLife.19760}.
}
\author{
Pirunthan Perampalam
}
