% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWeissmanScores.R
\name{loadWeissmanExample}
\alias{loadWeissmanExample}
\title{Get example data for CRISPRa or CRISPRi}
\usage{
loadWeissmanExample(modality = c("CRISPRa", "CRISPRi"))
}
\arguments{
\item{modality}{For which mode of perturbation should example data be
retrieved? Must be a \code{string} specifying either \code{CRISPRa} or
\code{CRISPRi}.}
}
\value{
\strong{loadWeissmanExample} returns a \code{list} containing two
\code{data.frame}s named \code{tssTable} and \code{sgrnaTable} that have
an example \code{tss_df} and \code{sgrna_df}, respectively, for the
selected CRISPR \code{modality}.
}
\description{
Retrieves example datasets for Cas9-based CRISPR activation
(CRISPRa) or  CRISPR inactivation (CRISPRi) modalities that can be used
in \code{getWeissmanScores}.
}
\examples{
\donttest{
ex.data <- loadWeissmanExample(modality="CRISPRa")
tssTable <- ex.data[["tssTable"]]
sgrnaTable <- ex.data[["sgrnaTable"]]
}

}
\author{
Pirunthan Perampalam
}
