% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mit.weights}
\alias{mit.weights}
\title{Mismatch position weights for the MIT off-target algorithm (Cas9)}
\format{
Numeric vector
}
\source{
\url{https://dx.doi.org/10.1038\%2Fnbt.2647}
}
\usage{
mit.weights
}
\description{
MIT off-target scoring weights for each nucleotide position, for 20-mer
    spacer sequences. Position 20 refers to the spacer position adjacent to
    the PAM sequence. A higher weight indicates greater mismatch tolerance 
    by the Cas9 nuclease.
}
\keyword{datasets}
