% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scoringMethodsInfo}
\alias{scoringMethodsInfo}
\title{data.frame detailing available scoring methods}
\format{
A data frame with 6 columns:
\describe{
  \item{method}{name of the scoring method}
  \item{nuclease}{nuclease compatible with the scoring method}
  \item{left}{upstream offset (relative to PAM site) to extract nucleotide
      sequence needed for scoring}
  \item{right}{downstream offset (relative to PAM site) to extract nucleotide
       sequence needed for scoring}
  \item{type}{type of the scoring algorithm (on-target or off-target)}
  \item{len}{length of the nucleotide sequence needed for scoring}
}
}
\usage{
scoringMethodsInfo
}
\description{
data.frame detailing available scoring methods with information needed
    to extract nucleotide sequences needed by each scoring algorithm.
}
\keyword{datasets}
