% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sgrnaExampleCrispra}
\alias{sgrnaExampleCrispra}
\title{Example CRISPRa gRNAs data.frame for the getCrispraiScores function}
\format{
A data.frame with 5 columns:
\describe{
  \item{grna_id}{String specifying gRNA unique identifier.}
  \item{tss_id}{String specifying the targeted TSS id.}
  \item{pam_site}{Genomic coordinate specifying the first nucleotide of 
       PAM sequence.}
  \item{strand}{Strand of the gRNA strand. Must be "+" or "-".}
  \item{spacer_19mer}{String specifying the nucleotide sequence of the 19mer
   spacer sequence.}
}
}
\usage{
data(sgrnaExampleCrispra)
}
\description{
Example CRISPRa gRNAs data.frame for the getCrispraiScores function. 
    The targeted TSSs are described in the object tssExampleCrispra.
}
\keyword{datasets}
