% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tssExampleCrispri}
\alias{tssExampleCrispri}
\title{Example TSS data.frame for the getCrispraiScores function}
\format{
A data.frame with 7 columns:
\describe{
  \item{tss_id}{String specifying the targeted TSS id.}
  \item{gene_symbol}{String specifying gene symbol.}
  \item{promoter}{String specifying the promoter suffix to add to the
  gene symbol columns to obtain the unique TSS id.}
  \item{transcripts}{Ensembl IDs of the targeted transcript.}
  \item{position}{Integer specifying genomic coordinate of the TSS.}
  \item{strand}{Strand of TSS. Must be "+" or "-".}
  \item{chr}{String specifying the chromosome name.}
}
}
\usage{
data(tssExampleCrispri)
}
\description{
Example TSS data for gRNAs stored in sgrnaExampleCrispri.
}
\keyword{datasets}
