% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCFDScores.R
\name{getCFDScores}
\alias{getCFDScores}
\title{Calculate CFD off-target specificity scores for CRISPR/Cas9}
\usage{
getCFDScores(spacers, protospacers)
}
\arguments{
\item{spacers}{Character vector of 20bp spacer sequences.}

\item{protospacers}{Character vector of 23bp protospacer sequences.}
}
\value{
\strong{getCFDScores} returns a data.frame with \code{spacer},
    \code{protospacer}, and \code{score} columns. The CFD score takes
    on a value between 0 and 1. For a given pair (on-target, off-target), 
    a higher CFD score indicates a higher likelihood for the Cas9 nuclease
    to cut at the off-target. Non-canonical PAM sequences are taken into
    account by the CFD algorithm.
}
\description{
Calculate cutting frequency determination (CFD) off-target
    specificity scores for CRISPR/Cas9.
}
\examples{
# Calculating MIT scores for two off-targets with respect to
# one spacer sequence:
spacer <- "AGGTGTAGTGTGTGTGATAA"
protospacer1 <- paste0("CGGTGTAGTGTGTGTGATAA", "AGG")
protospacer2 <- paste0("CGGTGTCGTGTGTGTGATAA", "CGG")
results <- getCFDScores(spacers=spacer,
    protospacers=c(protospacer1, protospacer2)
)

}
\references{
Doench, J., Fusi, N., Sullender, M. et al. Optimized sgRNA design to
    maximize activity and minimize off-target effects of CRISPR-Cas9.
    Nat Biotechnol 34, 184–191 (2016).
    \url{https://doi.org/10.1038/nbt.3437}.
}
\author{
Jean-Philippe Fortin
}
